Use message implementation for Send() and Recv(); reposition tests
